e-anamnesis-master/e-anamnesis-master/backend/requirements.txt
Django==3.1
djangorestframework==3.12
pyyaml
uritemplate
psycopg2
python-dotenv
